Сделал custom-link более универсальным
