--- src/components/custom-link/custom-link.test.jsx
@@ -1,14 +1,0 @@
-import renderer from "react-test-renderer";
-import CustomLink from './custom-link';
-
-it(`render correctly CustomLink componen`, () => {
-
-  const title = `Test`;
-  const handleLinkClick = () => {};
-
-  const tree = renderer
-    .create(<CustomLink title={title} onLinkClick={handleLinkClick}/>)
-    .toJSON();
-
-  expect(tree).toMatchSnapshot();
-});
